+++ mlrun/runtimes/__init__.py
@@ -21,7 +21,7 @@ from .kubejob import KubejobRuntime, KubeRuntimeHandler  # noqa
 from .local import HandlerRuntime, LocalRuntime  # noqa
 from .mpijob import MpiRuntimeV1Alpha1, MpiRuntimeV1, MpiV1RuntimeHandler, \
     MpiV1Alpha1RuntimeHandler  # noqa
-from .types import MPIJobCRDVersions
+from .constants import MPIJobCRDVersions
 from .nuclio import nuclio_init_hook
 from .serving import MLModelServer
 from .sparkjob import SparkRuntime, SparkRuntimeHandler  # noqa
